Changed all of the Logistic Regression Binary Classification calls to Multi-Nomial Classification Calls
Major was to change all of the Logistic Regression Binary Classification calls to Multi-Nomial calls.  This was a painful task but I wanted to make sure it was done correctly.  It is easy to do in the Graph Lab Turi API to create Multi-Nomial calls but the skLearn Logistic Regression API is painful.